requirements.txt
# Core
pytz==2018.4
Django==2.0.6
django-configurations==2.0
gunicorn==19.8.1
django-cors-headers==2.2.0

# For the persistence stores
psycopg2-binary==2.7.4
dj-database-url==0.5.0

# Model Tools
django-model-utils==3.1.2
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.8.2
Markdown==2.6.11
django-filter==1.1.0

# Developer Tools
ipdb==0.11
ipython==6.4.0
mkdocs==0.17.4
flake8==3.5.0

# Testing
mock==2.0.0
factory-boy==2.11.1
django-nose==1.4.5
nose-progressive==1.5.1
coverage==4.5.1

# Static and Media Storage
django-storages==1.6.6
boto3==1.7.39

# Workers
django-rq==1.1.0
rq-scheduler==0.8.3

# Web download
git+https://github.com/zTrix/webpage2html@master#egg=webpage2html
beautifulsoup4>=4.0.0
lxml>=3.4.4
requests>=2.5.2
termcolor>=1.1.0

# Personal libraries
-e git+https://github.com/nikitautiu/learnhtml.git@master#egg=learnhtml

# CLI tools
django-click==2.1.0